\name{ifa.init.pca}
\alias{ifa.init.pca}
\title{Internal function to initialize the IFA parameters to ordinary factor solution}
\description{
 An internal function to initialize the IFA parameters to ordinary factor solution. 
 This is not to be called by the user, since it is automatically called by the function \code{ifa.em} 
 if the option \code{init} is \code{NULL} (default).
}
\seealso{\code{\link{ifa.em}},\code{\link{ifa.init.random}}}
\keyword{multivariate}
